+++ opentelemetry-otlp/src/transform/mod.rs
@@ -1,5 +1,0 @@
-#[cfg(feature = "metrics")]
-mod metrics;
-
-#[cfg(all(feature = "grpc-tonic", feature = "metrics"))]
-pub(crate) use metrics::tonic::sink;